refactor: lower level JSX in 'pages' to separate 'organisms' component
now we can read 'pages' a lot easier. there is still some duplication,
however (ex. RecommendedCommunities, TrendingCommunities.tsx)
