--- src/components/organisms/Trending.tsx
@@ -1,13 +1,0 @@
-import React from 'react'
-
-import CommunityCard from '../molecules/CommunityCard'
-
-export default function Trending() {
-  return (
-    <CommunityCard
-      title="Rapid"
-      description="An aamzing super coolio awese app that auto manual repeat grid"
-      totalMembers="1235"
-    />
-  )
-}
